vertex_custom/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
scikit-learn==1.4.1.post1
xgboost==2.1.3
pandas==2.2.0
google-cloud-storage==2.14.0
